move lzstring to lib
