+++ src/lib/LZString/index.js
